refactor to student api logic and validations
