+++ src/main/java/com/tim/SampleWebApp/api/validators/student/StudentAddressValidator.java
@@ -2,7 +2,7 @@ package com.tim.SampleWebApp.api.validators.student;
 
 import org.apache.commons.lang3.StringUtils;
 
-import com.tim.SampleWebApp.api.validators.base.Validator;
+import com.tim.SampleWebApp.api.validators.Validator;
 import com.tim.SampleWebApp.common.CommonConstants.ApiMessages;
 import com.tim.SampleWebApp.error.Message;
 
@@ -17,12 +17,9 @@ public class StudentAddressValidator implements Validator<Object, Message> {
 		String str = (String) toValidate;
 		if (StringUtils.isBlank(str)) {
 			return m.constructFromEnumForField(ApiMessages.NULL_FIELD_VALUE, fieldName);
-		} else if (StringUtils.isAlphanumericSpace(str)) {
-			if (str.length() > fieldLength) {
-				m = m.constructFromEnumForField(ApiMessages.FIELD_LENGTH_TOO_LONG, fieldName);
-			}
-		} else {
-			m = m.constructFromEnumForField(ApiMessages.ALPHANUMERIC_ONLY, fieldName);
+		}
+		if (str.length() > fieldLength) {
+			m = m.constructFromEnumForField(ApiMessages.FIELD_LENGTH_TOO_LONG, fieldName);
 		}
 
 		return m;
