+++ src/main/java/com/tim/SampleWebApp/api/validators/student/StudentEmailValidator.java
@@ -4,7 +4,7 @@ import java.util.regex.Pattern;
 
 import org.apache.commons.lang3.StringUtils;
 
-import com.tim.SampleWebApp.api.validators.base.Validator;
+import com.tim.SampleWebApp.api.validators.Validator;
 import com.tim.SampleWebApp.common.CommonConstants.ApiMessages;
 import com.tim.SampleWebApp.error.Message;
